requirements.txt
absl-py==0.14.1
argon2-cffi==21.1.0
asgiref==3.4.1
astunparse==1.6.3
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
clang==5.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.8
django-cors-headers==3.10.0
django-crispy-forms==1.13.0
django-import-export==2.6.1
django-rest-swagger==2.2.0
djangorestframework==3.12.4
drf-yasg==1.20.0
emoji==1.6.1
entrypoints==0.3
et-xmlfile==1.1.0
flatbuffers==1.12
gast==0.4.0
google-api-core==2.1.1
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-cloud-vision==2.5.0
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.34.1
h5py==3.1.0
idna==3.3
inflection==0.5.1
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
jsonschema==4.1.2
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.8.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.6.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lxml==4.6.3
Markdown==3.3.4
MarkupPy==1.14
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
numpy==1.19.5
oauthlib==3.1.1
odfpy==1.4.1
openapi-codec==1.3.2
opencv-python==4.5.3.56
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.21
proto-plus==1.19.5
protobuf==3.18.1
py-hanspell @ git+https://github.com/ssut/py-hanspell.git@8e993cf46f97f9d665c15633a0fc78ac1b727713
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.10.0
pykospacing @ git+https://github.com/haven-jeon/PyKoSpacing.git@2e303d3c2b31efcd256bdb614d0856add7c87f3f
pyparsing==2.4.7
pyrsistent==0.18.0
pytesseract==0.3.8
python-dateutil==2.8.2
pytz==2021.3
pywin32==302
pywinpty==1.1.4
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
regex==2021.10.8
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
scikit-learn==1.0
scipy==1.7.1
Send2Trash==1.8.0
simplejson==3.17.5
six==1.15.0
sklearn==0.0
sqlparse==0.4.2
tablib==3.0.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.1
tensorflow-estimator==2.5.0
tensorflow-hub==0.12.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tornado==6.1
traitlets==5.1.0
typing-extensions==3.7.4.3
uritemplate==4.1.1
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
